# services/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/dashboard/statistics": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get dashboard statistics for the current organization",
        "operationId": "DashboardController_getStatistics",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dashboard statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatisticsResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Atlas AI API",
    "description": "Enterprise AI Platform API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DashboardStatisticsResponseDto": {
        "type": "object",
        "properties": {
          "workspacesCount": {
            "type": "integer",
            "description": "Number of workspaces in the organization",
            "example": 5
          },
          "organizationsCount": {
            "type": "integer",
            "description": "Number of organizations the user belongs to",
            "example": 2
          },
          "projectsCount": {
            "type": "integer",
            "description": "Number of active projects in the organization",
            "example": 12
          },
          "activeUsersCount": {
            "type": "integer",
            "description": "Number of active members in the organization",
            "example": 8
          }
        },
        "required": [
          "workspacesCount",
          "organizationsCount",
          "projectsCount",
          "activeUsersCount"
        ]
      },
...
    }
  }
}